Windows bring-up.
